+++ CHANGELOG.md
@@ -1,6 +1,30 @@
-# nymphemeral's changelog
+### nymphemeral 1.3.0, released 2015-03-03
 
-New on version 1.2.1:
+- Create client module ([#5f1ce14])
+- Modify the GUI to be a layer between the user and the client ([#5cd3fe1], [#3d093ff],
+[#7e50bc8], [#dde11dd], [#7d411ee], [#dabe7ed])
+
+[#5f1ce14]: https://github.com/felipedau/nymphemeral/commit/5f1ce140ffb812f71df3ca292c250a1d5ea092ee
+[#5cd3fe1]: https://github.com/felipedau/nymphemeral/commit/5cd3fe15b26c542565e9aed51e8ea8b38c1b1726
+[#3d093ff]: https://github.com/felipedau/nymphemeral/commit/3d093ff35f858e3791373d4551a0f255278e4742
+[#7e50bc8]: https://github.com/felipedau/nymphemeral/commit/7e50bc824b3e1e7c0e2080cce2996134f4df0195
+[#dde11dd]: https://github.com/felipedau/nymphemeral/commit/dde11dde600e2ecfd83fc0f7b48af9f8430fbda0
+[#7d411ee]: https://github.com/felipedau/nymphemeral/commit/7d411eea2dd40f1b0e4a573aced1116cf38574be
+[#dabe7ed]: https://github.com/felipedau/nymphemeral/commit/dabe7ede0dce40200dc369bfc81c0a9d7071cabb
+
+### nymphemeral 1.2.3, released 2015-02-14
+
+- Remove dependency links processing from `pip install` ([#0d18709])
+
+[#0d18709]: https://github.com/felipedau/nymphemeral/commit/0d1870964138b9607de7730c33750fb38099f0e3
+
+### nymphemeral 1.2.2.1, released 2014-11-14
+
+- Remember the output method being used ([#b5e6f90])
+
+[#b5e6f90]: https://github.com/felipedau/nymphemeral/commit/b5e6f9052c618e96c293fdc6196277dd0eff2412
+
+### nymphemeral 1.2.1, release 2014-11-10
 
 - Append date to the title of the messages in the 'inbox' ([#77f0fe2])
 - Encrypt hSub passphrases ([#4a93903], [#963098e])